css/css-borders-4/parsing/border-radius-right-invalid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Borders 4 Test: Parsing 'border-radius-right' with invalid values</title>
  <link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
  <link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-sizing">
  <meta name="assert"
    content="'border-radius-right' supports only the syntax <length-percentage [0,∞]>{1,2} [ / <length-percentage [0,∞]>{1,2} ]?">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <script>
    test_invalid_value("border-radius-right", "auto");
    test_invalid_value("border-radius-right", "none");
    test_invalid_value("border-radius-right", "1px 2px 3px");
    test_invalid_value("border-radius-right", "-1px");
    test_invalid_value("border-radius-right", "1px -2px");
    test_invalid_value("border-radius-right", "1em /");
    test_invalid_value("border-radius-right", "1px / 2px 3px 4px");
    test_invalid_value("border-radius-right", "1px / 2px / 3px");
    test_invalid_value("border-radius-right", "4");
    test_invalid_value("border-radius-right", "4 / 5");
  </script>
</body>
</html>
